+++ parcels/particle.py
@@ -160,13 +160,13 @@ class ScipyParticle(_Particle):
     Additional Variables can be added via the :Class Variable: objects
     """
 
-    lon = Variable('lon', dtype=np.float32)
-    lat = Variable('lat', dtype=np.float32)
-    depth = Variable('depth', dtype=np.float32)
+    lon = Variable('lon', dtype=np.float64)
+    lat = Variable('lat', dtype=np.float64)
+    depth = Variable('depth', dtype=np.float64)
     time = Variable('time', dtype=np.float64)
     id = Variable('id', dtype=np.int32)
     fileid = Variable('fileid', dtype=np.int32, to_write=False)
-    dt = Variable('dt', dtype=np.float32, to_write=False)
+    dt = Variable('dt', dtype=np.float64, to_write=False)
     state = Variable('state', dtype=np.int32, initial=ErrorCode.Success, to_write=False)
 
     def __init__(self, lon, lat, fieldset, depth=0., time=0., cptr=None):
